Added local storage, username validation to server for put and delete requests.

+++ routes/ideas.js
@@ -44,33 +44,49 @@ router.post('/', async (req, res) => {
 // update idea
 router.put('/:id', async (req, res) => {
   try {
-    const updatedIdea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text,
-          tag: req.body.tag
-        }
-      },
-      { new: true }
-    );
-    res.json({ success: true, data: updatedIdea });
+    const idea = await Idea.findById(req.params.id);
+
+    // match the usernames
+    if (idea.username === req.body.username) {
+      const updatedIdea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text,
+            tag: req.body.tag
+          }
+        },
+        { new: true }
+      );
+      return res.json({ success: true, data: updatedIdea });
+    } else {
+      // username doesnt match
+      res.status(403).json({ success: false, error: 'You are not authorized to update this resource!' });
+    };
   } catch (error) {
     console.log(error);
-    res.status(500).json({ success: false, error: 'Something went wrong.' })
-  }
+    res.status(500).json({ success: false, error: 'Something went wrong.' });
+  };
 });
 
 // delete an idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
-    res.json({ success: true, data: {} });
+    const idea = await Idea.findById(req.params.id);
+
+    // match the usernames
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
+      return res.json({ success: true, data: {} });
+    } else {
+    // usernames dont match
+    res.status(403).json({ success: false, error: 'You are not authorized to delete this resource!' })
+    };
   } catch (error) {
     console.log(error);
     res.status(500).json({ success: false, error: 'Something went wrong.' })
   }
-})
+});
 
 module.exports = router;
